move send() and get_template_context(self) from triggers classes to trigger base class
